Update the release branch with 2021.1 content

+++ demos/interactive_face_detection_demo/detectors.hpp
@@ -23,8 +23,6 @@
 #include <samples/common.hpp>
 #include <samples/slog.hpp>
 
-#include <ie_iextension.h>
-
 #include <opencv2/opencv.hpp>
 
 // -------------------------Generic routines for detection networks-------------------------------------------------
@@ -74,8 +72,8 @@ struct FaceDetection : BaseDetection {
     int enquedFrames;
     float width;
     float height;
-    float network_input_width;
-    float network_input_height;
+    size_t network_input_width;
+    size_t network_input_height;
     float bb_enlarge_coefficient;
     float bb_dx_coefficient;
     float bb_dy_coefficient;
