+++ demos/smart_classroom_demo/src/action_detector.cpp
@@ -177,8 +177,8 @@ ActionDetection::ParseBBoxRecord(const float* data, bool inverse) const {
 inline ActionDetection::NormalizedBBox
 ActionDetection::GeneratePriorBox(int pos, int step, const cv::Size2f& anchor,
                                   const cv::Size& blob_size) const {
-    const float row = pos / blob_size.width;
-    const float col = pos % blob_size.width;
+    const int row = pos / blob_size.width;
+    const int col = pos % blob_size.width;
 
     const float center_x = (col + 0.5f) * static_cast<float>(step);
     const float center_y = (row + 0.5f) * static_cast<float>(step);
